Game: StigmaService — equip/prereq validation + Item.Slot int->long fix
StigmaService (EquipStigma/UnequipStigma: slot-unlock by level/quest, class restrict,
shard consume, advanced-stigma <require_skill> prerequisites, unequip-blocked-by-
dependent guard, skill grant/remove), ItemTemplate stigma prereq + class-restrict
parse, CM_EQUIP_ITEM stigma path -> service, enter-world validate+auto-unequip pass,
real SM_CUBE_UPDATE.StigmaSlots count. FIXES a latent bug: Item.Slot int->long (stigma
slots use bit 32-52, overflowed int32) + V44 widens player_items/warehouse slot to
BIGINT + updated the affected packets. Membership-slot bypass stays // note:.

+++ AionLightning.NET/AionLightning.Game/Dao/ItemDaoImpl.cs
@@ -34,7 +34,7 @@ public sealed class ItemDaoImpl : IItemDao
                 UniqueId       = reader.GetInt64(0),
                 ItemId         = reader.GetInt32(1),
                 Count          = reader.GetInt64(2),
-                Slot           = reader.GetInt32(3),
+                Slot           = reader.GetInt64(3),
                 StorageType    = reader.GetByte(4),
                 EnchantLevel   = reader.GetByte(5),
                 GodStoneItemId = reader.GetInt32(6),
@@ -108,7 +108,7 @@ public sealed class ItemDaoImpl : IItemDao
                 UniqueId     = reader.GetInt64(0),
                 ItemId       = reader.GetInt32(1),
                 Count        = reader.GetInt64(2),
-                Slot         = reader.GetInt32(3),
+                Slot         = reader.GetInt64(3),
                 EnchantLevel = reader.GetByte(4),
                 StorageType  = 2,
             });
