+++ AionLightning.NET/AionLightning.Game/Network/Aion/ServerPackets/SM_VIEW_PLAYER_DETAILS.cs
@@ -45,7 +45,7 @@ public sealed class SM_VIEW_PLAYER_DETAILS : AionServerPacket
         w.WriteH(0);
         w.WriteD(0);
 
-        w.WriteH(item.Slot);
+        w.WriteH((int)item.Slot);
         w.WriteC(0); // cloth flag
 
         if (item.EnchantLevel > 0)
